requirements.txt
fastapi==0.128.0
uvicorn==0.40.0
SQLAlchemy==2.0.45
pandas==2.3.3
pydantic==2.12.5
pydantic-settings==2.1.0
python-jose==3.5.0
passlib==1.7.4
bcrypt==4.0.1
python-multipart==0.0.21
reportlab==4.0.7
psycopg2-binary==2.9.10
PyPDF2==3.0.1
openpyxl==3.1.2
python-dotenv==1.0.0
